Modify datatype of ref_relative_dist
Modified datatype of ref_relative_dist to match
that of display_order_hint

+++ av1/encoder/encoder.h
@@ -1041,7 +1041,7 @@ typedef struct AV1_COMP {
   int deltaq_used;
 
   // Indicates the true relative distance of ref frame w.r.t. current frame
-  int8_t ref_relative_dist[INTER_REFS_PER_FRAME];
+  int ref_relative_dist[INTER_REFS_PER_FRAME];
 
   // Indicate nearest references w.r.t. current frame in past and future
   int8_t nearest_past_ref;
